+++ CHANGELOG.md
@@ -1,3 +1,22 @@
 ## 0.0.1
 
 Initial release!
+
+## 0.0.0-placeholder
+
+Mistake in versioning. Forgive.
+
+## 1.0.0
+
+### StickWidget is now Stick
+
+...and .show property replaced with .index
+
+### CrawlStickerSurface.
+
+Now there is new properties:
+1. sticksCount - count of sticks. It is neede to generate GlobalKeys
+2. stickBuilder - builder for indicator and worm sticks.
+3. selected - index of selected stick.
+
+The .duration and .curve prefixed with 'animation'.